Rework as a custom resource
